refactor: Rename isTrackingDisabled into areAnalyticsDisabled
