+++ lib/utils/analytics.js
@@ -8,7 +8,7 @@ const BbPromise = require('bluebird');
 const pLimit = require('p-limit');
 const fetch = require('node-fetch');
 const fse = require('fs-extra');
-const isTrackingDisabled = require('./isTrackingDisabled');
+const areAnalyticsDisabled = require('./areAnalyticsDisabled');
 const log = require('./log/serverlessLog');
 
 const timestampWeekBefore = Date.now() - 1000 * 60 * 60 * 24 * 7;
@@ -99,7 +99,7 @@ function request(type, event, { id, timeout } = {}) {
 function track(type, event, options = {}) {
   return BbPromise.try(() => {
     const isForced = options && options.isForced;
-    if (isTrackingDisabled && !isForced) return null;
+    if (areAnalyticsDisabled && !isForced) return null;
     if (!cacheDirPath) return request(type, event);
     const id = uuid();
     return BbPromise.all([
@@ -143,7 +143,7 @@ function sendPending(options = {}) {
     if (options.serverlessExecutionSpan) {
       options.serverlessExecutionSpan.then(markServerlessRunEnd, markServerlessRunEnd);
     }
-    if (isTrackingDisabled && !isForced) return null;
+    if (areAnalyticsDisabled && !isForced) return null;
     if (!cacheDirPath) return null;
 
     const limit = pLimit(3);
