+++ lib/utils/analytics.test.js
@@ -35,7 +35,7 @@ describe('analytics', () => {
   };
   before(() => {
     ({ track, sendPending } = proxyquire('./analytics.js', {
-      './isTrackingDisabled': false,
+      './areAnalyticsDisabled': false,
       'node-fetch': url => {
         usedUrl = url;
         ++pendingRequests;
