+++ lib/utils/areAnalyticsDisabled.js
